+++ R/2-proprCall.R
@@ -157,5 +157,55 @@ simplify <- function(object){
 updateCutoffs.propr <- function(object, cutoff = seq(.05, .95, .3)){
 
   if(identical(object@permutes, list(NULL))) stop("Permutation testing is disabled.")
-  stop("Uh oh! This function is not yet implemented.")
+
+  # Let NA cutoff skip function
+  if(identical(cutoff, NA)) return(object)
+
+  # Set up FDR cutoff table
+  FDR <- as.data.frame(matrix(0, nrow = length(cutoff), ncol = 4))
+  colnames(FDR) <- c("cutoff", "randcounts", "truecounts", "FDR")
+  FDR$cutoff <- cutoff
+  p <- length(object@permutes)
+
+  # Calculate propr for each permutation -- NOTE: `select` and `subset` disable permutation testing
+  for(k in 1:p){
+
+    numTicks <- progress(k, p, numTicks)
+
+    # Calculate propr exactly based on @metric, @ivar, and @alpha
+    ct.k <- object@permutes[[k]]
+    pr.k <- suppressMessages(
+      propr(ct.k, object@metric, ivar = object@ivar, alpha = object@alpha))
+    pkt <- pr.k@propr$propr
+
+    # Find number of permuted theta less than cutoff
+    for(cut in 1:nrow(FDR)){ # randcounts as cumsum
+
+      # Count positives as rho > cutoff, cor > cutoff, phi < cutoff, phs < cutoff
+      if(object@metric == "rho" | object@metric == "cor"){
+        FDR[cut, "randcounts"] <- FDR[cut, "randcounts"] + sum(pkt > FDR[cut, "cutoff"])
+      }else{ # phi & phs
+        FDR[cut, "randcounts"] <- FDR[cut, "randcounts"] + sum(pkt < FDR[cut, "cutoff"])
+      }
+    }
+  }
+
+  # Calculate FDR based on real and permuted tallys
+  FDR$randcounts <- FDR$randcounts / p # randcounts as mean
+  for(cut in 1:nrow(FDR)){
+
+    # Count positives as rho > cutoff, cor > cutoff, phi < cutoff, phs < cutoff
+    if(object@metric == "rho" | object@metric == "cor"){
+      FDR[cut, "truecounts"] <- sum(object@propr$propr > FDR[cut, "cutoff"])
+    }else{ # phi & phs
+      FDR[cut, "truecounts"] <- sum(object@propr$propr < FDR[cut, "cutoff"])
+    }
+
+    FDR[cut, "FDR"] <- FDR[cut, "randcounts"] / FDR[cut, "truecounts"]
+  }
+
+  # Initialize @fdr
+  object@fdr <- FDR
+
+  return(object)
 }